+++ lib/rspec/mocks.rb
@@ -17,7 +17,6 @@ RSpec::Support.define_optimized_require_for_rspec(:mocks) { |f| require_relative
   order_group
   error_generator
   space
-  extensions/marshal
   mutate_const
   targets
   syntax
@@ -114,9 +113,10 @@ module RSpec
 
     # To speed up boot time a bit, delay loading optional or rarely
     # used features until their first use.
-    autoload :AnyInstance, "rspec/mocks/any_instance"
-    autoload :ExpectChain, "rspec/mocks/message_chain"
-    autoload :StubChain,   "rspec/mocks/message_chain"
+    autoload :AnyInstance,      "rspec/mocks/any_instance"
+    autoload :ExpectChain,      "rspec/mocks/message_chain"
+    autoload :StubChain,        "rspec/mocks/message_chain"
+    autoload :MarshalExtension, "rspec/mocks/marshal_extension"
 
     # Namespace for mock-related matchers.
     module Matchers